+++ inc/define.php
@@ -448,10 +448,6 @@ $CFG_GLPI["globalsearch_types"]           = ['Computer', 'Contact', 'Contract',
 $CFG_GLPI["number_format"]  = 0;
 $CFG_GLPI["decimal_number"] = 2;
 
-// Default debug options : may be locally overriden
-$CFG_GLPI["debug_sql"] = $CFG_GLPI["debug_vars"] = $CFG_GLPI["debug_lang"] = 1;
-
-
 // User Prefs fields which override $CFG_GLPI config
 $CFG_GLPI['user_pref_field'] = ['backcreated', 'csv_delimiter', 'date_format',
     'default_requesttypes_id', 'display_count_on_home',