+++ src/Api/API.php
@@ -52,6 +52,7 @@ use Document;
 use Dropdown;
 use Glpi\Api\HL\Router;
 use Glpi\DBAL\QueryExpression;
+use Glpi\Debug\Profile;
 use Glpi\Search\Provider\SQLProvider;
 use Glpi\Search\SearchOption;
 use Glpi\Toolbox\MarkdownRenderer;
@@ -1572,9 +1573,6 @@ abstract class API
      */
     protected function searchItems($itemtype, $params = [])
     {
-        /** @var array $DEBUG_SQL */
-        global $DEBUG_SQL;
-
         $itemtype = $this->handleDepreciation($itemtype);
 
        // check rights
@@ -1673,16 +1671,13 @@ abstract class API
        // force reset
         $params['reset'] = 'reset';
 
-       // force logging sql queries
-        $_SESSION['glpi_use_mode'] = Session::DEBUG_MODE;
-
        // call Core Search method
         $rawdata = Search::getDatas($itemtype, $params, $params['forcedisplay']);
 
        // probably a sql error
         if (!isset($rawdata['data']) || count($rawdata['data']) === 0) {
             $this->returnError(
-                "Unexpected SQL Error : " . array_splice($DEBUG_SQL['errors'], -2)[0],
+                'An internal error occured while trying to fetch the data.',
                 500,
                 "ERROR_SQL",
                 false
@@ -2352,9 +2347,6 @@ abstract class API
      */
     private function getGlpiLastMessage()
     {
-        /** @var array $DEBUG_SQL */
-        global $DEBUG_SQL;
-
         $all_messages             = [];
 
         $messages_after_redirect  = [];
@@ -2376,11 +2368,8 @@ abstract class API
         }
 
        // get sql errors
-        if (
-            count($all_messages) <= 0
-            && ($DEBUG_SQL['errors'] ?? null) !== null
-        ) {
-            $all_messages = $DEBUG_SQL['errors'];
+        if (count($all_messages) <= 0) {
+            $all_messages = Profile::getCurrent()->getSQLErrors();
         }
 
         if (!end($all_messages)) {